+++ @remirror/react-ssr/src/__tests__/remirror-ssr.server.spec.tsx
@@ -17,7 +17,7 @@ beforeEach(() => {
 });
 
 test('should render the ssr component', () => {
-  manager.init({ getState: () => state, getPortalContainer: () => new NodeViewPortalContainer() });
+  manager.init({ getState: () => state, getPortals: () => new NodeViewPortalContainer() });
   const state = manager.createState({ content: initialJson });
 
   const htmlString = renderToString(